Move build functions into directories
